PMM-2990: Drop singletons generated by cobra. * Easier to write parallel tests. * A bit better separation of app from cobra commands.
