+++ app/server/server.go
@@ -1,4 +1,4 @@
-package cmd
+package server
 
 import (
 	"context"
