+++ pmm-agent.go
@@ -2,8 +2,23 @@
 
 package main
 
-import "github.com/percona/pmm-agent/cmd"
+import (
+	"fmt"
+	"os"
+
+	"github.com/percona/pmm-agent/app"
+	"github.com/percona/pmm-agent/cmd"
+)
 
 func main() {
-	cmd.Execute()
+	app := &app.App{}
+	if err := app.Config.Read(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+
+	if err := cmd.New(app).Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
 }
